Django Loader: New config: CELERY_DB_REUSE_MAX, (max number of times to reuse the db connection)
The default is to use a new connection for every task.
We would very much like to reuse the connection, but a safe number of reuses
is not known, and we don't have any way to handle the errors that might
happen, which may even be database dependent.

See: http://groups.google.com/group/django-users/browse_thread/thread/78200863d0c07c6d/