Added support for router classes (like the django multidb routers)
New setting: CELERY_ROUTES

This is a single, or a list of routers to traverse when
sending tasks. Dicts in this list converts to a "celery.routes.MapRoute"
instance.

Example:

    >>> CELERY_ROUTES = {"celery.ping": "default",
                          "mytasks.add": "cpu-bound",
                          "video.encode": {
                            "queue": "video",
                            "exchange": "media"
                            "routing_key": "media.video.encode"}}

    >>> CELERY_ROUTES = ("myapp.tasks.Router",
                         {"celery.ping": "default})

Where myapp.tasks.Router could be:

    class Router(object):

        def route_for_task(self, task, task_id=None, args=None, kwargs=None):
            if task == "celery.ping":
                return "default"

route_for_task may return a string or a dict. A string then means
it's a queue name in CELERY_QUEUES, a dict means it's a custom route.

When sending tasks, the routers are consulted in order. The first
router that doesn't return None is the route to use. The message options
is then merged with the found route settings, where the routers settings
have priority.

Example if apply_async has these arguments::

    Task.apply_async(immediate=False, exchange="video",
routing_key="video.compress")

and a router returns::

    {"immediate": True,
     "exchange": "urgent"}

the final message options will be::

    immediate=True, exchange="urgent", routing_key="video.compress"

(and any default message options defined in the Task class)

Closes #79