cPickle is broken in Python <= 2.5.
It unsafely and incorrectly uses relative instead of absolute imports,
so e.g.:
      exceptions.KeyError
becomes:
      celery.exceptions.KeyError

Your best choice is to upgrade to Python 2.6,
as while the pure pickle version has worse performance,
it is the only safe option for older Python versions.

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.