No longer have to type PYTHONPATH=. to use celeryconfig in current dir.
This is accomplished by the default loader ensuring that the current directory
is in sys.path when loading the config module. sys.path is reset to its
original state after loading.

Adding cwd to sys.path without the user knowing may be a security issue,
as this means someone can drop a python module in the users directory that
executes arbitrary commands. This was the original reason not to do this,
but if done *only when loading the config module*, this means that this
behvavior will only apply to the modules imported in the config module,
which I think is a good compromise (certainly better than just explictly
setting PYTHONPATH=. anyway)