Configuration must appear after '--':
    celeryd -l info -- broker.host=foo.com

The configuration keys are case-insensitive and adhers to the following
format: {NAMESPACE}_{KEYNAME}. The keys are automatically converted to
uppercase and .'s are replaced with _'s (e.g. ``broker.host`` becomes
``BROKER_HOST``).
Namespace can be one of: BROKER, CELERY, CELERYD,
                         CELERYBEAT, CELERYMON,
                         EMAIL.

If the key starts with "." or "_" the default namespace is used,
for celeryd this is "CELERYD", for celerybeat it's "CELERYBEAT", ..., and
so on.

    celeryd -l info -- .concurrency=8 \
                       .pool=celery.concurrency.threads.TaskPool \
                       celery_create_missing_queues=false

becomes:

    {"CELERYD_CONCURRENCY": 8,
     "CELERYD_POOL": "celery.concurrency.threads.TaskPool",
     "CELERY_CREATE_MISSING_QUEUES": False}