Fixed fast growing memory leak introduced in master branch (not present in 2.1.x or previous versions).
We use a modified task publisher that only declares exchanges it has not
already declared previously in the same process.  To keep track of this
we use a global set.  With kombu `Publisher.exchange` is a
`kombu.entity.Exchange` instance instead, so it would always add the
exchange to this set.

Thanks to PiotrSikora! Closes #271.