Do a better job cleaning up state between tests
Two problems are fixed here. The first dealt with the faking of a PyPy
version; unfortunately on my system this caused us to completely skip
the database backend tests because we leaked the pypy_version_info
attribute into later tests, and these tests are conditional on this
attribute not being defined. Fix this by deleting the attribute
completely if it was not defined when entering the context manager.

The second one is more subtle- because request is a Context() object,
which is a thread local, setting the chord object to a value persisted
it for all remaining tests. Clear out the context at the end of the
test, and set it to a more realistic value.