Improve how exc_info is passed into loggers.
This adjusts all loggers to pass exc_info=True rather than
sys.exc_info, with the goal to standardize how logging happens
within Celery. The logging framework itself (as well as things like
Sentry) understand that this means "capture the exc info if youre
going to use it". So in certain cases (maybe loggings disabled for
whatever crazy reason), this could potentially be an optimization.

It also benefits that the loggers take care of grabbing the exc_info
and Celery doesnt need to deal with sys.exc_info/cleanup unless
it explicitly wants to.