Change CLA text to something, well, really more ambiguous
We need to find a text that everyone can agree on.
Obviously one can't just change the text of legal documents
so consider the previous document void.

From now on there is a new policy where people need
to sign themselves into the CONTRIBUTORS.txt file, so that
they're aware that this means the project
wants the rights to do, well, anything we want, with the
contributions they maek, and that they may be receive a request to sign
an official contributor agreement at some point in the future.