Fix 2to3 race which prevents installation on Py3k
setup.py correctly identifies that it needs to run 2to3 but, before it does so,
attempts to import celery.app which results in an AttributeError due to celery's
usage of iteritems().

	Downloading/unpacking Celery
	  Running setup.py egg_info for package Celery
		Traceback (most recent call last):
		  File "<string>", line 14, in <module>
		  File "/Users/jed/[snip]/build/Celery/setup.py", line 16, in <module>
			import celery.app
		  File "celery/__init__.py", line 38, in <module>
			VERSION=VERSION, SERIES=SERIES, VERSION_BANNER=VERSION_BANNER,
		  File "celery/__compat__.py", line 164, in recreate_module
			origins = get_origins(by_module)
		  File "celery/__compat__.py", line 196, in get_origins
			for module, items in defs.iteritems():
		AttributeError: 'dict' object has no attribute 'iteritems'

This is a temporary emergency fix which allows Celery to install successfully on
Pythons where 2to3 is required but better handling of this should be examined.