Adds app.amqp.autoexchange to specify the exchange for automatic queues.
The main purpose for this addition is to be able to use the AMQP 'anonymous
exchange':

    app = Celery()
    app.amqp.autoqueue = lambda name: None

The anon exchange can be used to bypass routing and send messages to a queue
directly by name.  The name of the queue is specified in the `routing_key`.
Using the anon exchange will speed up message publish.