add note about Django 1.6 transaction changes to userguide
See #2472.

Django 1.6 introduced a change to the tranasction model, switching to autocommit
by default and deprecating much of the old transaction API, with plans to
remove it completely in 1.8.

This commit adds a note to the userguide section that discusses race conditions
involving database transactions, pointing out the ramifications of the
example code in Django 1.6+