Fixed sphinx-build check in Makefile (#4728)
If sphinx-build is not installed, the Makefile returns the cryptic
'recipe commences before first target.  Stop.' when building the docs.
The cause is a tab resulting in a syntax error. This was a bug in the
Makefile generated by Sphinx, which has since been fixed:
https://github.com/sphinx-doc/sphinx/issues/2521

I've applied the same fix, which is removing the 'user-friendly'
message. The resulting error message is 'sphinx-build: Command not
found', which ought to be clear enough.