Allow GroupResult.join timeout to be configurable in celery.chord_unlock (#5348)
* Allow GroupResult.join timeout to be configurable in celery.chord_unlock

Previously the timeout passed down to GroupResult.join in celery.chord_unlock was hardcoded to 3.0. This introduces the new configuration option result_chord_join_timeout which allows users to set this value themselves.

* Add configuration documentation for result_chord_join_timeout

* Reduce length of line with 80+ characters

* Allow GroupResult.join timeout to be configurable in celery.chord_unlock

Previously the timeout passed down to GroupResult.join in celery.chord_unlock was hardcoded to 3.0. This introduces the new configuration option result_chord_join_timeout which allows users to set this value themselves.

* Add configuration documentation for result_chord_join_timeout

* Reduce length of line with 80+ characters

* Added unit tests for result_chord_join_timeout

* Added self as a contributor

* Updated documentation

* Remove code duplication in introduced test cases

* Fix typo

* flake8 fixes