fix: Error propagation and errback calling for group-like signatures (#6746)
* fix: Use chord kwarg over request in group.apply

* fix: Propagate errors from failed chain tasks

Fixes #6220

Co-authored-by: Maximilian Friedersdorff <max@friedersdorff.com>
Co-authored-by: maybe-sybr <58414429+maybe-sybr@users.noreply.github.com>

* fix: Ensure all subtasks of a group get errbacks

Giving a linked task callback to the 0th task in a group is fine, but
for errbacks it's not an appropriate choice since any task in the group
could fail. This ensures that if any task other than the 0th one fails,
the errback will be called. This opens the possibility for an errback to
be called more than once when linked to a group, but generally we expect
that they should be design to be idempotent so no warning is issued for
the changed behaviour.

* test: Add tests for child error propagation

* test: Add regression tests for group errback dupes

These tests simply encode the currently expected behaviour where
errbacks linked to a group will be called once for each failed task, as
well as the consequences for chords which turn their header into a group
if it is not one already.

* doc: Add extra docs for canvas call/errback usage

Co-authored-by: Crawford, Jordan <Jordan.Crawford@msci.com>
Co-authored-by: Maximilian Friedersdorff <max@friedersdorff.com>