fix: Preserve call/errbacks of replaced tasks (#6770)
* style: Remove unused var from canvas unit tests

* test: Check task ID re-freeze on replacement

* refac: Remove duped task ID preservation logic

* test: Rework canvas call/errback integration tests

This change modifies a bunch of the tests to use unique keys for the
`redis_echo` and `redis_count` tasks which are used to validate that
callbacks and errbacks are made. We also introduce helper functions for
validating that messages/counts are seen to reduce duplicate code.

* fix: Preserve call/errbacks of replaced tasks

Fixes #6441

* fix: Ensure replacement tasks get the group index

This change adds some tests to ensure that when a task is replaced, it
runs as expected. This exposed a bug where the group index of a task
would be lost when replaced with a chain since chains would not pass
their `group_index` option down to the final task when applied. This
manifested as the results of chords being mis-ordered on the redis
backend since the group index would default to `+inf`. Other backends
may have had similar issues.